restructured task holder
