+++ src/lib/kombi/TaskHolder/Dispatcher/Local/Local.py
@@ -4,7 +4,7 @@ import subprocess
 import tempfile
 import threading
 from ..Dispatcher import Dispatcher, DispatcherError
-from ...ProcessExecution import ProcessExecution
+from ....ProcessExecution import ProcessExecution
 
 class LocalExecutionError(DispatcherError):
     """Local Execution Error."""
