+++ src/lib/kombi/TaskHolder/Loader/__init__.py
@@ -1,4 +1,4 @@
-from .TaskHolderLoader import TaskHolderLoader, TaskHolderLoaderError, TaskHolderLoaderInvalidConfigError, TaskHolderLoaderNotRegisteredError
+from .Loader import Loader, LoaderError, LoaderInvalidConfigError, LoaderNotRegisteredError
 from .PythonLoader import PythonLoader, PythonLoaderContentError
 from .JsonLoader import JsonLoader
 from .YamlLoader import YamlLoader
