+++ src/lib/kombi/TaskHolder/__init__.py
@@ -1,0 +1,3 @@
+from .TaskHolder import TaskHolder, TaskHolderError, TaskHolderInvalidVarNameError
+from . import Loader
+from . import Dispatcher
